Change session messages structure for reviews and posts
The session success messages have been altered for both the create and update methods on the reviews controller as well as the post controller. Instead of setting session.success directly, a messages object has been created within the session for more structured and potentially multipurpose messaging. By using this structure, we can elegantly manage multiple types of session messages if we need to in the future. The messages object is associated with the session to ensure that it's destroyed after the session ends.