Move toggling review form functionality from EJS to JS file
Moved the JavaScript that handles toggling of the review form from 'show.ejs' to 'show.js' in order to centralize JS interactions and leave EJS files more focused on templating and presentation logic. This helps in improving code organization and separation of concerns. 'show.js' is more appropriate for script handling.